Creation du deplacement du personnage, quelque bug de sprite
Les sprites ne continuent pas lorque le perso avance.

+++ src/fr/iutvalence/projetS2/java/groupe7/Position.java
@@ -12,12 +12,12 @@ public class Position
 	/**
 	 * Type position X (horizontal)
 	 */
-	public static int X ;
+	public static float X ;
 	
 	/**
-	 * type poistion Y (vertical)
+	 * type position Y (vertical)
 	 */
-	public static int Y ;
+	public static float Y ;
 	
 	
 }
